Fixed bug with CONSTANT_Class items referring to array types (which they can). Fixed InGen bug regarding SIPUSH opcode -- operand was being created   incorrectly.  Three cheers for self-checking code.

+++ ClassFile/Bytecode/InGen.java
@@ -112,7 +112,7 @@ public class InGen extends Instr {
       // Short
     case Op.SIPUSH:
       operands = new Operand[] { 
-	new OpConstant(new Short((short)u2(code,pc+1)), HClass.Byte)
+	new OpConstant(new Short((short)u2(code,pc+1)), HClass.Short)
       };
       break;
       // Integer constant.
